requirements.txt
Django==1.9.4
django-timezones==0.2
Pillow==3.1.1
pytz==2016.6.1
requests==2.12.4
wheel==0.24.0
gunicorn==19.6.0
